Move roi methods from views.py to ROI_utils with tests
