+++ components/tools/OmeroPy/test/unit/test_rois.py
@@ -1,0 +1,48 @@
+#!/usr/bin/env python
+# -*- coding: utf-8 -*-
+
+#
+# Copyright (C) 2015 University of Dundee & Open Microscopy Environment.
+# All rights reserved.
+#
+# This program is free software; you can redistribute it and/or modify
+# it under the terms of the GNU General Public License as published by
+# the Free Software Foundation; either version 2 of the License, or
+# (at your option) any later version.
+#
+# This program is distributed in the hope that it will be useful,
+# but WITHOUT ANY WARRANTY; without even the implied warranty of
+# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+# GNU General Public License for more details.
+#
+# You should have received a copy of the GNU General Public License along
+# with this program; if not, write to the Free Software Foundation, Inc.,
+# 51 Franklin Street, Fifth Floor, Boston, MA 02110-1301 USA.
+
+"""
+Simple tests of various ROI utilities
+"""
+
+from omero.util.ROI_utils import pointsStringToXYlist, xyListToBbox
+
+
+class TestRoiUtils(object):
+
+    def testOldFormat(self):
+        xyList = pointsStringToXYlist((
+            "points[1,2, 3,4, 5,6] "
+        ))
+        assert xyList == [(1, 2), (3, 4), (5, 6)]
+        return xyList
+
+    def testNewFormat(self):
+        xyList = pointsStringToXYlist((
+            "1,2 3,4 5,6"
+        ))
+        assert xyList == [(1, 2), (3, 4), (5, 6)]
+        return xyList
+
+    def testBBox(self):
+        xyList = self.testOldFormat()
+        bbox = xyListToBbox(xyList)
+        assert bbox == (1, 2, 4, 4)
